Added methods to multipy by Killo / Mega / Giga etc

+++ Camalot.Common/Camalot.Common/Extensions/Decimal.cs
@@ -11,7 +11,7 @@ namespace Camalot.Common.Extensions {
 		/// </summary>
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
-		public static bool IsEven(this float value) {
+		public static bool IsEven(this decimal value) {
 			return (value % 2) == 0;
 		}
 
@@ -20,10 +20,9 @@ namespace Camalot.Common.Extensions {
 		/// </summary>
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
-		public static bool IsOdd(this float value) {
+		public static bool IsOdd(this decimal value) {
 			return (value % 2) != 0;
 		}
-
 
 		/// <summary>
 		/// Multiplies the value by 2^10
@@ -31,7 +30,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="00deb6720cc278be8ca9" />
-		public static float Kilo(this float value) {
+		public static decimal Kilo(this decimal value) {
 			return value * 1024L;
 		}
 
@@ -41,7 +40,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="00deb6720cc278be8ca9" />
-		public static float K(this float value) {
+		public static decimal K(this decimal value) {
 			return value * 1000L;
 		}
 
@@ -53,7 +52,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="d01704cf490ec59fc1e9" />	
-		public static float Mega(this float value) {
+		public static decimal Mega(this decimal value) {
 			return value.Kilo() * 1024L;
 		}
 
@@ -62,8 +61,8 @@ namespace Camalot.Common.Extensions {
 		/// </summary>
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
-		/// <gist id="d01704cf490ec59fc1e9" />	
-		public static float M(this float value) {
+		/// <gist id="d01704cf490ec59fc1e9" />
+		public static decimal M(this decimal value) {
 			return value.K() * 1000L;
 
 		}
@@ -75,7 +74,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="f1b33489ed80d3f8bc0f" />
-		public static float Giga(this float value) {
+		public static decimal Giga(this decimal value) {
 			return value.Mega() * 1024L;
 		}
 
@@ -85,7 +84,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="f1b33489ed80d3f8bc0f" />
-		public static float B(this float value) {
+		public static decimal B(this decimal value) {
 			return value.M() * 1000;
 		}
 
@@ -95,7 +94,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="1061db9b00620852af88" />
-		public static float T(this float value) {
+		public static decimal T(this decimal value) {
 			return value.B() * 1000L;
 		}
 
@@ -105,7 +104,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="1061db9b00620852af88" />
-		public static float Tera(this float value) {
+		public static decimal Tera(this decimal value) {
 			return value.Giga() * 1024L;
 		}
 
@@ -115,7 +114,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="c9dc04638a360990984a" />
-		public static float Peta(this float value) {
+		public static decimal Peta(this decimal value) {
 			return value.Tera() * 1024L;
 		}
 
@@ -125,7 +124,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="c9dc04638a360990984a" />
-		public static float P(this float value) {
+		public static decimal P(this decimal value) {
 			return value.T() * 1000;
 		}
 
@@ -135,7 +134,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="2d53fbbd5442e8b731a0" />
-		public static float Exa(this float value) {
+		public static decimal Exa(this decimal value) {
 			return value.Peta() * 1024L;
 		}
 
@@ -145,7 +144,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="2d53fbbd5442e8b731a0" />
-		public static float E(this float value) {
+		public static decimal E(this decimal value) {
 			return value.P() * 1000;
 		}
 
@@ -155,7 +154,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="28b00f2ec3e2c83ef173" />
-		public static float Zetta(this float value) {
+		public static decimal Zetta(this decimal value) {
 			return value.Exa() * 1024L;
 		}
 
@@ -165,7 +164,7 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="28b00f2ec3e2c83ef173" />
-		public static float Z(this float value) {
+		public static decimal Z(this decimal value) {
 			return value.E() * 1000;
 		}
 
@@ -176,8 +175,8 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="749614ebc92dc68211b7" />
-		public static float Yotta(this float value) {
-			return value.Zetta() * 1024F;
+		public static decimal Yotta(this decimal value) {
+			return value.Zetta() * 1024;
 		}
 
 		/// <summary>
@@ -186,9 +185,8 @@ namespace Camalot.Common.Extensions {
 		/// <param name="value">The value.</param>
 		/// <returns></returns>
 		/// <gist id="749614ebc92dc68211b7" />
-		public static float Y(this float value) {
-			return value.Z() * 1000F;
+		public static decimal Y(this decimal value) {
+			return value.Z() * 1000;
 		}
-
 	}
 }
